pyproject.toml
[project]
name = "bakken-sweet-spot"
version = "0.2.0"
description = "Sweet-spot and production analysis for North Dakota Bakken wells"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "linearmodels>=7.0",
    "matplotlib>=3.10.9",
    "numpy>=2.4.4",
    "pandas>=3.0.3",
    "pillow>=12.2.0",
    "pyyaml>=6.0",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.6",
]

[project.optional-dependencies]
map = [
    "contextily>=1.6.2",
    "geopandas>=1.0.1",
]
dev = [
    "openpyxl>=3.1.5",
    "pytest>=8.0",
    "ruff>=0.9",
]

[project.scripts]
bakken = "bakken_sweet_spot.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bakken_sweet_spot"]

[tool.uv]
package = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
